backend/api/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.5
tensorflow>=2.10.0
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.0.0
requests>=2.25.0
